+++ Proyectoprogbien/Cliente.h
@@ -13,7 +13,7 @@ protected:
     string apellido;
     int estado; //1 activo 0 de baja
     int antiguedad;
-    float saldo;
+    long double saldo;
     int membresia;
 public:
     Transaccion transacciones[100];
@@ -27,17 +27,17 @@ public:
     int getestado();
     void setantiguedad(int _antiguedad);
     int getantiguedad();
-    void setsaldo(float _saldo);
-    float getsaldo();
+    void setsaldo(long double _saldo);
+    long double getsaldo();
     void setmembresia (int _membresia);
     int getmembresia();
 
     //metodos especificos
     void imprimircliente();
-    void extraccion( float _monto, int _dia, int _mes, int _anio);
-    void deposito( float _monto, int _dia, int _mes, int _anio);
+    void extraccion( long double _monto, int _dia, int _mes, int _anio);
+    void deposito( long double _monto, int _dia, int _mes, int _anio);
 
     //Constructores
     Cliente();
-    Cliente(int _numcliente, string _nombre, string _apellido, int _estado, int _antiguedad, float _saldo, int _membresia);
+    Cliente(int _numcliente, string _nombre, string _apellido, int _estado, int _antiguedad, long double _saldo, int _membresia);
 };
